Refactoring of XML parsing
Moved low-level XML parsing to DomUtils, and moved DomUtils to com.vzome.xml.
