+++ core/src/main/java/com/vzome/core/editor/Snapshot.java
@@ -10,8 +10,8 @@ import org.w3c.dom.Element;
 
 import com.vzome.core.commands.Command.Failure;
 import com.vzome.core.commands.XmlSaveFormat;
-import com.vzome.core.math.DomUtils;
 import com.vzome.core.render.RenderedModel;
+import com.vzome.xml.DomUtils;
 
 public class Snapshot extends UndoableEdit
 {
