+++ core/src/main/java/com/vzome/core/edits/ColorManifestations.java
@@ -13,11 +13,11 @@ import com.vzome.core.commands.Command.Failure;
 import com.vzome.core.commands.XmlSaveFormat;
 import com.vzome.core.editor.ChangeManifestations;
 import com.vzome.core.editor.Selection;
-import com.vzome.core.math.DomUtils;
 import com.vzome.core.model.Manifestation;
 import com.vzome.core.model.RealizedModel;
 import com.vzome.core.render.Color;
 import com.vzome.core.render.RenderedManifestation;
+import com.vzome.xml.DomUtils;
 
 public class ColorManifestations extends ChangeManifestations
 {
