+++ core/src/main/java/com/vzome/core/edits/GroupSelection.java
@@ -13,7 +13,7 @@ import com.vzome.core.commands.XmlSaveFormat;
 import com.vzome.core.editor.EditorModel;
 import com.vzome.core.editor.Selection;
 import com.vzome.core.editor.UndoableEdit;
-import com.vzome.core.math.DomUtils;
+import com.vzome.xml.DomUtils;
 
 public class GroupSelection extends UndoableEdit
 {
