+++ core/src/main/java/com/vzome/core/edits/PolarZonohedron.java
@@ -22,12 +22,12 @@ import com.vzome.core.construction.TransformedSegment;
 import com.vzome.core.editor.ChangeManifestations;
 import com.vzome.core.editor.EditorModel;
 import com.vzome.core.editor.SymmetrySystem;
-import com.vzome.core.math.DomUtils;
 import com.vzome.core.math.symmetry.Axis;
 import com.vzome.core.math.symmetry.Permutation;
 import com.vzome.core.model.Connector;
 import com.vzome.core.model.Manifestation;
 import com.vzome.core.model.Strut;
+import com.vzome.xml.DomUtils;
 
 public class PolarZonohedron extends ChangeManifestations
 {
