+++ core/src/main/java/com/vzome/core/edits/SelectCollinear.java
@@ -14,9 +14,9 @@ import com.vzome.core.commands.Command.Failure;
 import com.vzome.core.editor.ChangeManifestations;
 import com.vzome.core.editor.EditorModel;
 import com.vzome.core.commands.XmlSaveFormat;
-import com.vzome.core.math.DomUtils;
 import com.vzome.core.model.Connector;
 import com.vzome.core.model.Strut;
+import com.vzome.xml.DomUtils;
 
 /**
  * @author David Hall
